Now it's working again & changed heartbeat-length to 43

+++ Game Server/src/de/gymdon/inf1315/game/packet/Packet.java
@@ -20,7 +20,7 @@ public abstract class Packet {
     public abstract void send() throws IOException;
     public abstract short getId();
 
-    public static Packet newPacket(int id, Client c) {
+    public static Packet newPacket(short id, Client c) {
 	if (packetTypes.containsKey(id))
 	    try {
 		return packetTypes.get(id).getConstructor(Client.class).newInstance(c);
